hashmap refactoring

+++ features/bspace.c
@@ -29,7 +29,7 @@ struct network *bspace_compute(struct network *net) {
 struct network *compute(struct network *net, bool comp) {
     /*** compute the maximum amount of contexts ***/
     int sam = maximum_state_amount(net);
-    int ct_amount = pow(sam, net->aut_amount) * net->lk_amount;
+    size_t ct_amount = pow(sam, net->aut_amount) * net->lk_amount;
     
     ct_hashmap = hashmap_create(ct_amount);
